Initial changes to return and modules structure
