+++ boa_engine/src/vm/opcode/control_flow/mod.rs
@@ -1,5 +1,8 @@
 pub(crate) mod r#break;
 pub(crate) mod catch;
+pub(crate) mod r#continue;
+pub(crate) mod r#return;
+pub(crate) mod throw;
 pub(crate) mod finally;
 pub(crate) mod labelled;
 pub(crate) mod r#try;
@@ -7,5 +10,8 @@ pub(crate) mod r#try;
 pub(crate) use catch::*;
 pub(crate) use finally::*;
 pub(crate) use labelled::*;
+pub(crate) use throw::*;
 pub(crate) use r#break::*;
+pub(crate) use r#continue::*;
 pub(crate) use r#try::*;
+pub(crate) use r#return::*;
