+++ boa_engine/src/vm/opcode/control_flow/throw.rs
